chore(browser): move test files to __tests__ directories
- Move agent-browser.test.ts, errors.test.ts, integration.test.ts, types.test.ts to __tests__/
- Move stagehand-browser.test.ts to __tests__/
- Update import paths to use ../
